Rename response as answer
