+++ app/routes/receive-message.js
@@ -19,8 +19,8 @@ const pausedMiddleware = require('../../lib/middleware/receive-message/template-
 const campaignMenuMiddleware = require('../../lib/middleware/receive-message/campaign-menu');
 const campaignKeywordMiddleware = require('../../lib/middleware/receive-message/campaign-keyword');
 const currentCampaignMiddleware = require('../../lib/middleware/receive-message/campaign-current');
-const parseAskSignupMiddleware = require('../../lib/middleware/receive-message/parse-ask-signup-response');
-const parseAskContinueMiddleware = require('../../lib/middleware/receive-message/parse-ask-continue-response');
+const parseAskSignupMiddleware = require('../../lib/middleware/receive-message/parse-ask-signup-answer');
+const parseAskContinueMiddleware = require('../../lib/middleware/receive-message/parse-ask-continue-answer');
 const askContinueMiddleware = require('../../lib/middleware/receive-message/template-ask-continue');
 const setCampaignMiddleware = require('../../lib/middleware/receive-message/conversation-set-campaign');
 const gambitReplyMiddleware = require('../../lib/middleware/receive-message/template-gambit');
