+++ lib/middleware/receive-message/parse-ask-continue-answer.js
