+++ lib/middleware/receive-message/parse-ask-signup-answer.js
